small short overflow fix in Postgresql.hpp tried to reactiviate parallel build on Unix (needed for OSC, so far no success)

+++ src/modules/database/postgresql/PostgreSQL.hpp
@@ -51,7 +51,7 @@ namespace db {
 
 static const char* POSTGRESQL_DB_CLASS_NAME = "PostgreSQL";
 
-static const unsigned short POSTGRESQL_MIN_DB_VERSION = 80400;
+static const unsigned int POSTGRESQL_MIN_DB_VERSION = 80400;
 static const unsigned short POSTGRESQL_MIN_PROTOCOL_VERSION = 3;
 
 /// PostgreSQL server connection configuration
